backend/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coverage==4.5.4
deprecation==2.0.7
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
janus==0.4.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mock==3.0.5
mysql-connector==2.2.9
packaging==19.2
postcodes-io-api==0.0.4
pyparsing==2.4.5
requests==2.22.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
